==> notebooks/00_workbook.ipynb <==
import src.config as config
import hopsworks

# connect to the project
project = hopsworks.login(
    project=config.HOPSWORKS_PROJECT_NAME,
    api_key_value=config.HOPSWORKS_API_KEY
)

# connect to the feature store
feature_store = project.get_feature_store()

# connect to the feature group
# feature_group = feature_store.get_or_create_feature_group(
#     name='model_predictions_feature_group',
#     version=7,
#     description="Predictions generate by our production model",
#     primary_key = ['pickup_location_id', 'pickup_ts'],
#     event_time='pickup_ts'
# )

feature_view = feature_store.get_feature_view(
    name="time_series_hourly_feature_view",
    version=7
)

ts_data = feature_view.get_batch_data(
    start_time=1655539200
)

# ts_data = feature_view.get_batch_data()

print(f'{ts_data.shape=}')

conn = hopsworks.Connection()
conn.close()

ts_data

import datetime
import time

timestamp = datetime.datetime.now().timestamp()

print(f'{timestamp=}')

current_time_seconds = int(time.time())
print(f"Seconds since Unix Epoch: {current_time_seconds}")


t='20180515102500'
d=datetime.datetime.strptime(t, "%Y%m%d%H%M%S")
print(d)
epoch = datetime.datetime.utcfromtimestamp(0)


def unix_time_millis(dt):
  return int((dt - epoch).total_seconds() * 1000)

unix_time_millis(d)
